Change workflow trigger type to run tests from external contributors
Workflows that use on: pull_request have no access to secrets from forks, even after approval, unless the repo is explicitly configured to “send secrets to workflows from pull requests.

Change to using pull_request_target, note that it does run in the context of the base repository, so it can access secrets. But this has its own security implications - which we are handling like in other repositories from the OSS Build team

Additional information:
https://stackoverflow.com/questions/76952023/how-to-make-github-actions-safely-access-secrets-for-prs-created-from-forks\?utm_source\=chatgpt.com

Example of CI job failing due to lack of access to secrets after we authorise the external pipeline to run:
https://github.com/astronomer/astro-provider-anyscale/pull/67