Change workflow trigger type to run tests from external contributors (#90)
This PR aims to allow us to authorise PRs from external contributors to
run integration tests using secrets set in Github actions in the
https://github.com/astronomer/astro-provider-anyscale repository.

Example of CI job failing due to lack of access to secrets after we
authorise the external pipeline to run:
https://github.com/astronomer/astro-provider-anyscale/pull/67

**About the change**

Workflows that use `on: pull_request` have no access to secrets from
forks, even after approval, unless the repo is explicitly configured to
"send secrets to workflows from pull requests" - which is not the case
here.

Change to using `pull_request_target`. Note that it runs in the context
of the base repository to access secrets. But this has its own security
implications, which we are handling like in other repositories from the
OSS Build team, such as https://github.com/astronomer/astronomer-cosmos.

Additional information:

https://stackoverflow.com/questions/76952023/how-to-make-github-actions-safely-access-secrets-for-prs-created-from-forks\?utm_source\=chatgpt.com